Removed VectorToRaster, moved last code to example
